requirements.txt
# Core ML Libraries
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0
datasets>=2.10.0

# Data Processing
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Configuration & Utilities  
PyYAML>=6.0
tqdm>=4.65.0
requests>=2.30.0

# Multilingual & Multimodal
sentence-transformers>=2.2.0
Pillow>=9.5.0
opencv-python>=4.7.0

# Dataset Downloads
gdown>=4.7.0
kaggle>=1.5.0

# Optional: Experiment Tracking
wandb>=0.15.0

# Optional: Performance
accelerate>=0.20.0
